backend-API-1/requirements.txt
Flask==2.2.2
Pillow==9.3.0
torch==1.12.0
torchvision==0.13.0
ultralytics==8.0.150
numpy==1.22.2
